refactor: extract SessionDiscoveryCoordinator from AppModel (Phase 2)
Move session discovery, persistence, and merge logic into a dedicated
SessionDiscoveryCoordinator. This extracts ~330 lines from AppModel
including startup discovery, Codex/Claude session restore, rollout
tracking, and all merge helpers. The coordinator communicates with
AppModel via closure-based state accessors, keeping SessionState.apply()
as the single source of truth.
